@@ -35,12 +35,13 @@
     ],
     "devDependencies": {
         "@aitodotai/json-stringify-pretty-compact": "^1.3.0",
-        "@playwright/test": "^1.43.0",
+        "@playwright/test": "^1.43.1",
         "@sveltejs/adapter-auto": "^3.2.0",
         "@sveltejs/adapter-static": "^3.0.1",
-        "@sveltejs/kit": "^2.5.5",
+        "@sveltejs/eslint-config": "^7.0.1",
+        "@sveltejs/kit": "^2.5.7",
         "@sveltejs/package": "^2.3.1",
-        "@sveltejs/vite-plugin-svelte": "^3.0.2",
+        "@sveltejs/vite-plugin-svelte": "^3.1.0",
         "@sveltepress/theme-default": "^2.0.2",
         "@sveltepress/twoslash": "^0.2.2",
         "@sveltepress/vite": "^0.33.2",
@@ -56,27 +57,27 @@
         "@types/d3-scale-chromatic": "^3.0.3",
         "@types/d3-shape": "^3.1.6",
         "@types/underscore": "^1.11.15",
-        "@typescript-eslint/eslint-plugin": "^7.6.0",
-        "@typescript-eslint/parser": "^7.6.0",
+        "@typescript-eslint/eslint-plugin": "^7.7.0",
+        "@typescript-eslint/parser": "^7.7.0",
         "d3-dsv": "^3.0.1",
         "d3-fetch": "^3.0.1",
         "d3-force": "^3.0.0",
-        "eslint": "^9.0.0",
+        "eslint": "^9.1.0",
         "eslint-config-prettier": "^9.1.0",
-        "eslint-plugin-svelte": "^2.36.0",
+        "eslint-plugin-svelte": "^2.37.0",
         "jsdom": "^24.0.0",
         "prettier": "^3.2.5",
-        "prettier-plugin-svelte": "^3.2.2",
+        "prettier-plugin-svelte": "^3.2.3",
         "remark-code-extra": "^1.0.1",
         "remark-code-frontmatter": "^1.0.0",
-        "sass": "^1.74.1",
+        "sass": "^1.75.0",
         "svelte-check": "^3.6.9",
         "svelte-highlight": "^7.6.0",
         "topojson-client": "^3.1.0",
         "tslib": "^2.6.2",
-        "typescript": "^5.4.4",
-        "vite": "^5.2.8",
-        "vitest": "^1.4.0"
+        "typescript": "^5.4.5",
+        "vite": "^5.2.10",
+        "vitest": "^1.5.0"
     },
     "svelte": "./dist/index.js",
     "types": "./dist/index.d.ts",
@@ -99,7 +100,7 @@
         "fast-equals": "^5.0.1",
         "merge-deep": "^3.0.3",
         "numeral": "^2.0.6",
-        "svelte": "5.0.0-next.96",
+        "svelte": "5.0.0-next.110",
         "underscore": "^1.13.6"
     }
 }
